requirements.txt
# Core dependencies
fastapi==0.128.8
uvicorn==0.39.0
python-dotenv==1.0.0

# AI/ML
openai==1.72.0
anthropic==0.84.0

# Vector Database
qdrant-client==1.16.1

# Data processing
requests==2.32.5
tqdm==4.67.3

# Zendesk API (if needed)
# zenpy==2.0.25
